[cppgrep] Fix project warnings

+++ test/cppgrep_all_test.cpp
@@ -8,5 +8,5 @@ int main()
 {
     const auto cpp_file = "\"" + (RESOURCES / "people.cpp").generic_string() + "\"";
     std::array argv { "cppgrep", "-cstfv", cpp_file.c_str() };
-    return static_cast<int>(klang::cppgrep::main(argv.size(), argv.data()));
+    return static_cast<int>(klang::cppgrep::main(static_cast<int>(argv.size()), argv.data()));
 }
